organize constants into separate file
